requirements.txt
Django
BeautifulSoup
Markdown
PIL
South
django-athumb
django-extensions
django-fabtastic
django-pagination
django-mediasync
django-taggit
django-typogrify
docutils
easy-thumbnails
fabric
gondor
pycrypto
pysolr
python-dateutil
psycopg2
smartypants
-e git+http://github.com/mikemeyer/django-shorturls.git#egg=django-shorturls
# -e git+http://github.com/toastdriven/django-haystack.git#egg=django-haystack
wsgiref
werkzeug
xapian-haystack
